refactor(ffi): domain directories, engine-core parts, ARCHITECTURE.md
Contributor-facing readability pass, move-not-rewrite:

- lib/src reorganized by domain: engine/ (protocol, host, core, frame
  driver, vsync), view/ (widget, gestures, ornaments), platform/
  (MapLibrePlatform adapter, offline API), io/ (HTTP provider, texture
  bridge, style resolution, JSON converters). Pure git mv; only import
  paths changed, the public barrel is identical.
- engine_core.dart shrinks from ~1500 lines to a ~170-line shell (state
  + frame pump); dispatch moves verbatim into same-library part files
  as extensions (commands, queries, offline, snapshots, session), so
  private access and execution semantics are untouched (Dart has no
  partial classes; extensions in parts are the idiomatic split).
- ARCHITECTURE.md: end-user/contributor tour in the style of the docs
  site's architecture page: mermaid flowchart of the stack, sequence
  diagram of one command end to end, state diagram of the frame loop,
  threading rules in plain language, and the 'add a new map API in 4
  steps' recipe. README points to it.

Gate: analyzer clean, gallery smoke on device (map renders, vsync
pulses active, idle park healthy).
