+++ maplibre_gl_native/lib/src/engine/frame_stats.dart
