+++ maplibre_gl_native/lib/src/engine/vsync_pulse.dart
