+++ maplibre_gl_native/lib/src/io/style_string_resolver.dart
