+++ maplibre_gl_native/lib/src/platform/ffi_platform_base.dart
